rename NetworkPlanner.py to network_planner.py
